requirements.txt
Django>=5.0.4,<5.1
djangorestframework>=3.15.1,<3.16
pillow>=10.3.0,<10.4
djangorestframework-simplejwt>=5.3.1,<5.4
django-cors-headers
coreapi
setuptools